+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [0.5.0]
 
+### Added
+
+- Sentry configuration for better exception tracking
+
 ### Modified
 
 - Exception handling in API
@@ -11,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Pydoc strings to make them uniform
 - Add unit tests related to `services` and `utils`
 - Replace `matplotlib` `set_tight_layout()` with `set_layout_engine` due to deprecation
+- Replace `config` by `Settings` class for better settings organization
 
 ## [0.4.4] - 11/06/2024